Fix an issue retrying migration failures
The existing code would retry a database migration if it failed. If
the failure was such that it didn't release the lock, the retry will
always fail. To make this worse, if the user didn't have Metabase set
to the WARN log level, they would not see the original exception, but
only the lock exception.

This commit tries to release the lock if a failure occurs. This will
cause the real exception to bubble up on subsequent retries rather
than complaining about the lock.